requirements-data-pipeline.txt
# Data Pipeline Requirements
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets API
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
